Refactor: Move spending info formatting to Utils.
